refactor: 合并 Bootstrap/FileBackend/CrashHook 为单一 LogSystem
之前为了启动期 stdlib-only 需求把日志系统裂成 3 模块 3 个 init 调用，是
增量副作用、不是架构必要。

收归后：
- 单文件 ModuleFolders/Log/LogSystem.py，单点 install() 入口
- 通过懒 import rich.text（缺失回落到正则）保住启动期纯 stdlib 特性
- AiNiee.py 顶 4 行 → 2 行，唯一调用 install() 一次
- formatter / filter / crash hooks / faulthandler / cleanup 全在一处
- 行为完全等价：保留 32 字符 API key 脱敏、exc_info redact、--windowed
  下 stdout/stderr devnull 兜底、5MB×5 rotation、30 天清理
